unassign check empty; prevent deletion of assigned author

+++ eBookStoreClient/Pages/Authors/Delete.cshtml.cs
@@ -3,9 +3,11 @@ using eBookStoreClient.Constants;
 using eBookStoreClient.Utilities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
 using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
 namespace eBookStoreClient.Pages.Authors
@@ -24,6 +26,8 @@ namespace eBookStoreClient.Pages.Authors
 
         [TempData]
         public int AuthorId { get; set; }
+
+        public string HasBookMessage { get; set; }
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
@@ -68,11 +72,23 @@ namespace eBookStoreClient.Pages.Authors
             try
             {
                 HttpClient httpClient = SessionHelper.GetHttpClient(HttpContext.Session, sessionStorage);
-                HttpResponseMessage response = await httpClient.DeleteAsync($"{Endpoints.Authors}/{AuthorId}");
+                HttpResponseMessage response = await httpClient.GetAsync($"{Endpoints.BookAuthors}?$filter=AuthorId eq {AuthorId}");
                 HttpContent content = response.Content;
                 if (response.StatusCode == HttpStatusCode.OK)
                 {
-                    return RedirectToPage(PageRoute.Authors);
+                    var bookAuthors = JsonSerializer.Deserialize<BookAuthors>(await content.ReadAsStringAsync(), SerializerOptions.CaseInsensitive).List;
+                    if (bookAuthors.Count > 0)
+                    {
+                        HasBookMessage = "Unable to delete author. This author is assigned to book(s).";
+                        return await OnGetAsync(AuthorId);
+                    }
+                    httpClient = SessionHelper.GetHttpClient(HttpContext.Session, sessionStorage);
+                    response = await httpClient.DeleteAsync($"{Endpoints.Authors}/{AuthorId}");
+                    content = response.Content;
+                    if (response.StatusCode == HttpStatusCode.OK)
+                    {
+                        return RedirectToPage(PageRoute.Authors);
+                    }
                 }
                 return await OnGetAsync(Author.AuthorId);
             }
@@ -81,5 +97,11 @@ namespace eBookStoreClient.Pages.Authors
             }
             return RedirectToPage(PageRoute.Authors);
         }
+    }
+
+    class BookAuthors
+    {
+        [JsonPropertyName("value")]
+        public List<BookAuthor> List { get; set; }
     }
 }
